requirements.txt
PyYAML>=3.0
nose>=1.0
flake8>=2.5
